Utilize `bunyan` to structure server and CLI logging.
Starting to replace our `console.log` based logging with `bunyan` to
help structure our logs across multiple levels, and in the future change
where logs are written based on options passed, and how Collapsify is
invoked.

For example, when invoked as a CLI tool, we can display logs in a human
readable fashion, but when invoked as part of a server, we can write the
logs in a machine-readable fashion to syslogd.

As a transition point, this commit migrates our current logging to using
`bunyan`, but continuing to output on stdout. The lack of an option
passed to "v" now defaults to outputting on log level "warn". "1" now
corresponds to "info" and "2" is now "debug".