feat(server): implement server routing
Collapsify previously assumed all requests were valid, and simply looked
at the "url" query parameter on the request, with no respect giving to
the path.

This adds routes to the HTTP server implementation, which I use to
enforce path and method restrictions.

I'm also using this oppurtunity to implement two routes for collapsing
a resource: the continued legacy "GET /?url=" and a new "POST
/v1/collapse" with a JSON body containing a "url" property. The GET
route continues to return an HTML response, whereas the new POST route
returns a JSON response.

The GET route is now deprecated.