refactor(http-client): use `bhttp` and `yakaa` for improved socket keepalive
The HTTP client was getting "socket hang up" errors, and long pauses in
responses, that seemed to be caused by an incorrect handling of open
sockets in a connection pool.

`yakaa` is an improved HTTP agent, specifically for keepalive sockets.
However, `needle` doesn't seem to be using it properly, so it was
replaced with `bhttp`.