Run cssnano in default preset config
The config options we have been passing to cssnano go unused, they are
not sent in a way that cssnano can accept config options.
What it does do is go looking on the filesystem for a config file by
searching the CWD and $HOME, if you tell it to use the `default` preset
it doesn't go looking for config files on the filesystem.