Change fetch interface for better web compatability
The `fetch` function is now expected to return an object with three
methods, one for the contentType, and two for the body in different
forms.
`getAsString` for collapsers that want the content as a string.
`getAsArray` for collapsers that want the raw binary data, for node
enviroments this can be satasified by returning a Buffer since they
extend UInt8Array, and in Web runtimes with Blob, they be be convered
into an Array easily.